Added pass through formatter for JSON log records (in particular LogStash records)
Added pass through formatter for JSON log records (in particular LogStash records). The primary purpose of adding these loggers is to allow other logging frameworks to format log messages and log to the console (System.out). The JBoss Log Manager will intercept messages written to the console (System.out), and thus attempt to format the already formatted messages. The pass through formatter performs a number of checks against a log record. If the conditions are met, the raw message (ExtLogRecord.getMessage()) is returned from the format method. If the conditions are not met, it defers formatting to a delegate ExtFormatter instance (in this case LogstashFormatter).